backend/_requirements.txt
asgiref==3.6.0
attrs==22.2.0
backend==0.2.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-extensions==3.2.3
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
idna==3.4
iniconfig==2.0.0
numpy==1.25.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
posts==0.2.2
pycparser==2.21
PyJWT==2.8.0
pytest==7.2.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==2.0.4
psycopg2-binary==2.9.3